third_party/blink/web_tests/external/wpt/screen-wake-lock/wakelock-disabled-by-feature-policy.https.html

<!DOCTYPE html>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/feature-policy/resources/featurepolicy.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script>
  "use strict";

  const same_origin_src =
    "/feature-policy/resources/feature-policy-screen-wakelock.html";
  const cross_origin_src =
    get_host_info().HTTPS_REMOTE_ORIGIN + same_origin_src;

  promise_test(t => {
    return promise_rejects_dom(t, "NotAllowedError", navigator.wakeLock.request("screen"));
  }, 'Feature-Policy header {"screen-wake-lock" : []} disallows the top-level document.');

  async_test(t => {
    test_feature_availability(
      'navigator.wakeLock.request("screen")',
      t,
      same_origin_src,
      expect_feature_unavailable_default
    );
  }, 'Feature-Policy header {"screen-wake-lock" : []} disallows same-origin iframes.');

  async_test(t => {
    test_feature_availability(
      'navigator.wakeLock.request("screen")',
      t,
      cross_origin_src,
      expect_feature_unavailable_default
    );
  }, 'Feature-Policy header {"screen-wake-lock" : []} disallows cross-origin iframes.');
</script>
</body>
